Add `BasicMemory` implementation of virtual memory
