+++ crates/execution/ab-riscv-act4-runner/res/abundance/abundance-rv32i-max/sail.json
@@ -179,7 +179,7 @@
         },
         "size": {
           "len": 64,
-          "value": "0x400000"
+          "value": "0x80000"
         },
         "attributes": {
           "mem_type": "MainMemory",
