--- crates/execution/ab-riscv-act4-runner/src/interpreter.rs
@@ -1,126 +1,0 @@
-use ab_riscv_interpreter::prelude::*;
-
-pub(crate) struct Act4Memory<const BASE_ADDR: u64, const SIZE: usize> {
-    data: Box<[u8; SIZE]>,
-}
-
-impl<const BASE_ADDR: u64, const SIZE: usize> VirtualMemory for Act4Memory<BASE_ADDR, SIZE> {
-    fn read<T>(&self, address: u64) -> Result<T, VirtualMemoryError>
-    where
-        T: BasicInt,
-    {
-        let offset = address
-            .checked_sub(BASE_ADDR)
-            .ok_or(VirtualMemoryError::OutOfBoundsRead { address })?;
-
-        if offset.saturating_add(size_of::<T>() as u64) > self.data.len() as u64 {
-            return Err(VirtualMemoryError::OutOfBoundsRead { address });
-        }
-
-        // SAFETY: Only reading basic integers from initialized memory
-        unsafe {
-            Ok(self
-                .data
-                .as_ptr()
-                .cast::<T>()
-                .byte_add(offset as usize)
-                .read_unaligned())
-        }
-    }
-
-    unsafe fn read_unchecked<T>(&self, address: u64) -> T
-    where
-        T: BasicInt,
-    {
-        // SAFETY: Guaranteed by function contract
-        unsafe {
-            let offset = address.unchecked_sub(BASE_ADDR) as usize;
-            self.data
-                .as_ptr()
-                .cast::<T>()
-                .byte_add(offset)
-                .read_unaligned()
-        }
-    }
-
-    fn read_slice(&self, address: u64, len: u32) -> Result<&[u8], VirtualMemoryError> {
-        let offset = address
-            .checked_sub(BASE_ADDR)
-            .ok_or(VirtualMemoryError::OutOfBoundsRead { address })?;
-
-        if offset > self.data.len() as u64 {
-            return Err(VirtualMemoryError::OutOfBoundsRead { address });
-        }
-
-        self.data
-            .get(offset as usize..)
-            .and_then(|data| data.get(..len as usize))
-            .ok_or(VirtualMemoryError::OutOfBoundsRead { address })
-    }
-
-    fn read_slice_up_to(&self, address: u64, len: u32) -> &[u8] {
-        let Some(offset) = address.checked_sub(BASE_ADDR) else {
-            return &[];
-        };
-
-        if offset > self.data.len() as u64 {
-            return &[];
-        }
-
-        let remaining = self.data.get(offset as usize..).unwrap_or_default();
-        remaining.get(..len as usize).unwrap_or(remaining)
-    }
-
-    fn write<T>(&mut self, address: u64, value: T) -> Result<(), VirtualMemoryError>
-    where
-        T: BasicInt,
-    {
-        let offset = address
-            .checked_sub(BASE_ADDR)
-            .ok_or(VirtualMemoryError::OutOfBoundsWrite { address })?;
-
-        if offset.saturating_add(size_of::<T>() as u64) > self.data.len() as u64 {
-            return Err(VirtualMemoryError::OutOfBoundsWrite { address });
-        }
-
-        // SAFETY: Only writing basic integers to initialized memory
-        unsafe {
-            self.data
-                .as_mut_ptr()
-                .cast::<T>()
-                .byte_add(offset as usize)
-                .write_unaligned(value);
-        }
-
-        Ok(())
-    }
-
-    fn write_slice(&mut self, address: u64, data: &[u8]) -> Result<(), VirtualMemoryError> {
-        let offset = address
-            .checked_sub(BASE_ADDR)
-            .ok_or(VirtualMemoryError::OutOfBoundsWrite { address })?;
-
-        if offset > self.data.len() as u64 {
-            return Err(VirtualMemoryError::OutOfBoundsWrite { address });
-        }
-
-        let len = data.len();
-        self.data
-            .get_mut(offset as usize..)
-            .and_then(|data| data.get_mut(..len))
-            .ok_or(VirtualMemoryError::OutOfBoundsWrite { address })?
-            .copy_from_slice(data);
-
-        Ok(())
-    }
-}
-
-impl<const BASE_ADDR: u64, const SIZE: usize> Act4Memory<BASE_ADDR, SIZE> {
-    pub(crate) fn new() -> Self {
-        Self {
-            // TODO: Should have been just `::new()`, but https://github.com/rust-lang/rust/issues/53827
-            // SAFETY: Data structure filled with zeroes is a valid invariant
-            data: unsafe { Box::new_zeroed().assume_init() },
-        }
-    }
-}
