+++ crates/execution/ab-riscv-interpreter/src/basic.rs
@@ -4,8 +4,9 @@
 mod tests;
 
 use crate::{
-    Address, CustomErrorPlaceholder, ExecutionError, FetchInstructionResult, InstructionFetcher,
-    ProgramCounter, ProgramCounterError, RegisterFile, SystemInstructionHandler, VirtualMemory,
+    Address, BasicInt, CustomErrorPlaceholder, ExecutionError, FetchInstructionResult,
+    InstructionFetcher, ProgramCounter, ProgramCounterError, RegisterFile,
+    SystemInstructionHandler, VirtualMemory, VirtualMemoryError,
 };
 use ab_riscv_primitives::prelude::*;
 use core::hint::cold_path;
@@ -125,6 +126,189 @@ pub struct BasicInterpreterState<Regs, ExtState, Memory, IF, InstructionHandler>
     pub instruction_fetcher: IF,
     /// System instruction handler
     pub system_instruction_handler: InstructionHandler,
+}
+
+/// Basic memory implementation.
+///
+/// Flat structure, no rwx protections, no alignment requirements. It uses stack, so for larger
+/// allocation it'll need to be boxed (zero-initialized is fine) or a custom implementation to be
+/// used.
+///
+/// This implementation is intentionally basic and correct, but not the most performant. It is
+/// possible to have a more efficient implementation that skips certain checks by placing additional
+/// constraints on the program.
+///
+/// This works for simpler cases, while a more sophisticated implementation might prevent certain
+/// memory from being writable, supporting actual virtual memory with dynamically allocated memory
+/// pages, etc.
+#[derive(Debug, Copy, Clone)]
+#[repr(align(16))]
+pub struct BasicMemory<const BASE_ADDR: u64, const SIZE: usize> {
+    data: [u8; SIZE],
+}
+
+impl<const BASE_ADDR: u64, const SIZE: usize> VirtualMemory for BasicMemory<BASE_ADDR, SIZE> {
+    #[inline(always)]
+    fn read<T>(&self, address: u64) -> Result<T, VirtualMemoryError>
+    where
+        T: BasicInt,
+    {
+        let Some(offset) = address.checked_sub(BASE_ADDR) else {
+            cold_path();
+            return Err(VirtualMemoryError::OutOfBoundsRead { address });
+        };
+
+        if offset.saturating_add(size_of::<T>() as u64) > self.data.len() as u64 {
+            cold_path();
+            return Err(VirtualMemoryError::OutOfBoundsRead { address });
+        }
+
+        // SAFETY: Only reading basic integers from initialized memory
+        unsafe {
+            Ok(self
+                .data
+                .as_ptr()
+                .cast::<T>()
+                .byte_add(offset as usize)
+                .read_unaligned())
+        }
+    }
+
+    #[inline(always)]
+    unsafe fn read_unchecked<T>(&self, address: u64) -> T
+    where
+        T: BasicInt,
+    {
+        // SAFETY: Guaranteed by function contract
+        unsafe {
+            let offset = address.unchecked_sub(BASE_ADDR) as usize;
+            self.data
+                .as_ptr()
+                .cast::<T>()
+                .byte_add(offset)
+                .read_unaligned()
+        }
+    }
+
+    fn read_slice(&self, address: u64, len: u32) -> Result<&[u8], VirtualMemoryError> {
+        let Some(offset) = address.checked_sub(BASE_ADDR) else {
+            cold_path();
+            return Err(VirtualMemoryError::OutOfBoundsRead { address });
+        };
+
+        if offset > self.data.len() as u64 {
+            cold_path();
+            return Err(VirtualMemoryError::OutOfBoundsRead { address });
+        }
+
+        self.data
+            .get(offset as usize..)
+            .and_then(|data| data.get(..len as usize))
+            .ok_or(VirtualMemoryError::OutOfBoundsRead { address })
+    }
+
+    fn read_slice_up_to(&self, address: u64, len: u32) -> &[u8] {
+        let Some(offset) = address.checked_sub(BASE_ADDR) else {
+            cold_path();
+            return &[];
+        };
+
+        if offset > self.data.len() as u64 {
+            cold_path();
+            return &[];
+        }
+
+        let remaining = self.data.get(offset as usize..).unwrap_or_default();
+        remaining.get(..len as usize).unwrap_or(remaining)
+    }
+
+    #[inline(always)]
+    fn write<T>(&mut self, address: u64, value: T) -> Result<(), VirtualMemoryError>
+    where
+        T: BasicInt,
+    {
+        let Some(offset) = address.checked_sub(BASE_ADDR) else {
+            cold_path();
+            return Err(VirtualMemoryError::OutOfBoundsWrite { address });
+        };
+
+        if offset.saturating_add(size_of::<T>() as u64) > self.data.len() as u64 {
+            cold_path();
+            return Err(VirtualMemoryError::OutOfBoundsWrite { address });
+        }
+
+        // SAFETY: Only writing basic integers to initialized memory
+        unsafe {
+            self.data
+                .as_mut_ptr()
+                .cast::<T>()
+                .byte_add(offset as usize)
+                .write_unaligned(value);
+        }
+
+        Ok(())
+    }
+
+    fn write_slice(&mut self, address: u64, data: &[u8]) -> Result<(), VirtualMemoryError> {
+        let Some(offset) = address.checked_sub(BASE_ADDR) else {
+            cold_path();
+            return Err(VirtualMemoryError::OutOfBoundsWrite { address });
+        };
+
+        if offset > self.data.len() as u64 {
+            cold_path();
+            return Err(VirtualMemoryError::OutOfBoundsWrite { address });
+        }
+
+        let len = data.len();
+        let Some(target_data) = self
+            .data
+            .get_mut(offset as usize..)
+            .and_then(|data| data.get_mut(..len))
+        else {
+            cold_path();
+            return Err(VirtualMemoryError::OutOfBoundsWrite { address });
+        };
+
+        target_data.copy_from_slice(data);
+
+        Ok(())
+    }
+}
+
+impl<const BASE_ADDR: u64, const SIZE: usize> Default for BasicMemory<BASE_ADDR, SIZE> {
+    #[inline(always)]
+    fn default() -> Self {
+        Self { data: [0; _] }
+    }
+}
+
+impl<const BASE_ADDR: u64, const SIZE: usize> BasicMemory<BASE_ADDR, SIZE> {
+    /// Get a mutable slice of memory.
+    ///
+    /// This is primarily useful for setting up the program and should not be used beyond that.
+    pub fn get_mut_bytes(
+        &mut self,
+        address: u64,
+        size: usize,
+    ) -> Result<&mut [u8], VirtualMemoryError> {
+        let Some(offset) = address.checked_sub(BASE_ADDR) else {
+            cold_path();
+            return Err(VirtualMemoryError::OutOfBoundsRead { address });
+        };
+        let offset = offset as usize;
+
+        let Some(slice) = self
+            .data
+            .get_mut(offset..)
+            .and_then(|data| data.get_mut(..size))
+        else {
+            cold_path();
+            return Err(VirtualMemoryError::OutOfBoundsRead { address });
+        };
+
+        Ok(slice)
+    }
 }
 
 /// Basic instruction fetcher implementation.
